refactor(node-agent): change workload-manager file structure
