+++ node-agent/workload-manager/src/lib.rs
@@ -1,1 +1,2 @@
-pub mod workload_manager;
+pub mod container;
+pub mod namespace;
