+++ node-agent/workload-manager/src/namespace/client.rs
