--- node-agent/workload-manager/src/workload_manager/mod.rs
@@ -1,2 +1,0 @@
-pub mod container;
-pub mod namespace;
